move functionality to hooks/helpers
