+++ src/helpers/formatValue/index.ts
@@ -1,1 +1,2 @@
 export * from './formatEventFields';
+export * from './formatDefinitionsForDisplay';
