+++ src/hooks/contract/index.ts
@@ -3,3 +3,4 @@ export * from './useGetContractFiles';
 export * from './useGetContractEvents';
 export * from './useGetContractDocs';
 export * from './useGetContractTypings';
+export * from './useGetContractEndpointCount';
